+++ common/src/main/java/earth/terrarium/adastra/common/entities/vehicles/Rover.java
@@ -274,8 +274,7 @@ public class Rover extends Vehicle implements PlayerRideable, RadioHolder {
 
     public void consumeFuel() {
         if (level().isClientSide() || tickCount % 5 != 0) return;
-        var toExtract = FluidHooks.newFluidHolder(fluidContainer.getFluids().get(0).getFluid(), FluidHooks.buckets(0.001), null);
-        fluidContainer.extractFluid(toExtract, false);
+        fluidContainer.extractFluid(fluidContainer.getFluids().get(0).copyWithAmount(FluidHooks.buckets(0.001)), false);
     }
 
     public boolean hasEnoughFuel() {